A better way to declare a float

+++ src/ParkViewProtector.java
@@ -27,7 +27,7 @@ public class ParkViewProtector extends Canvas
 	public static final Color COLOR_BG_2	= Color.white;
 	public static final Color COLOR_TEXT_2	= Color.black;
 	
-	public static final Color STATS_BAR_BG	= new Color((float) 0, (float) 0, (float) 0, (float) 0.5);
+	public static final Color STATS_BAR_BG	= new Color(0F, 0F, 0F, 0.5F);
 	public static final Color STATS_BAR_FG	= Color.white;
 	public static final Color STATS_BAR_HP	= new Color(255, 0, 255);
 	public static final Color STATS_BAR_TP	= new Color(0, 255, 0);
